feat(machines): Add button set on machine listing page
