+++ src/app/machines/views/MachineList/MachineList.tsx
@@ -1,37 +1,42 @@
-import { useCallback, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 
 import type { ValueOf } from "@canonical/react-components";
 import { useDispatch, useSelector } from "react-redux";
 import { useStorageState } from "react-storage-hooks";
 
 import ErrorsNotification from "./ErrorsNotification";
-import MachineListControls from "./MachineListControls";
 import MachineListTable from "./MachineListTable";
 import { DEFAULTS } from "./MachineListTable/constants";
 
 import VaultNotification from "app/base/components/VaultNotification";
 import { useWindowTitle } from "app/base/hooks";
-import type { SetSearchFilter, SortDirection } from "app/base/types";
+import type { SortDirection } from "app/base/types";
 import { actions as controllerActions } from "app/store/controller";
 import { actions as generalActions } from "app/store/general";
 import { actions as machineActions } from "app/store/machine";
 import machineSelectors from "app/store/machine/selectors";
-import { FetchGroupKey } from "app/store/machine/types";
+import type { FetchGroupKey } from "app/store/machine/types";
 import { FilterMachines } from "app/store/machine/utils";
 import { useFetchMachinesWithGroupingUpdates } from "app/store/machine/utils/hooks";
 
 type Props = {
+  grouping: FetchGroupKey | null;
+  hiddenColumns: string[];
+  hiddenGroups: (string | null)[];
   headerFormOpen?: boolean;
   searchFilter: string;
-  setSearchFilter: SetSearchFilter;
+  setHiddenGroups: (groups: (string | null)[]) => void;
 };
 
 const DEFAULT_PAGE_SIZE = DEFAULTS.pageSize;
 
 const MachineList = ({
+  grouping,
+  hiddenColumns,
+  hiddenGroups,
   headerFormOpen,
   searchFilter,
-  setSearchFilter,
+  setHiddenGroups,
 }: Props): JSX.Element => {
   useWindowTitle("Machines");
   const dispatch = useDispatch();
@@ -43,49 +48,6 @@ const MachineList = ({
   const [sortDirection, setSortDirection] = useState<
     ValueOf<typeof SortDirection>
   >(DEFAULTS.sortDirection);
-  const [storedGrouping, setStoredGrouping] =
-    useStorageState<FetchGroupKey | null>(
-      localStorage,
-      "grouping",
-      DEFAULTS.grouping
-    );
-  // fallback to "None" if the stored grouping is not valid
-  const grouping: FetchGroupKey =
-    typeof storedGrouping === "string" &&
-    Object.values(FetchGroupKey).includes(storedGrouping)
-      ? storedGrouping
-      : DEFAULTS.grouping;
-  const [hiddenColumns, setHiddenColumns] = useStorageState<string[]>(
-    localStorage,
-    "machineListHiddenColumns",
-    []
-  );
-  const handleSetGrouping = useCallback(
-    (group: FetchGroupKey | null) => {
-      setStoredGrouping(group);
-      // clear selected machines on grouping change
-      // we cannot reliably preserve the selected state for individual machines
-      // as we are only fetching information about a group from the back-end
-      dispatch(machineActions.setSelectedMachines(null));
-    },
-    [setStoredGrouping, dispatch]
-  );
-  const handleSetSearchFilter = useCallback(
-    (filter: string) => {
-      setSearchFilter(filter);
-      // clear selected machines on filters change
-      // we cannot reliably preserve the selected state for groups of machines
-      // as we are only fetching information about a group from the back-end
-      // and the contents of a group may change when different filters are applied
-      dispatch(machineActions.setSelectedMachines(null));
-    },
-    [dispatch, setSearchFilter]
-  );
-  const [hiddenGroups, setHiddenGroups] = useStorageState<(string | null)[]>(
-    localStorage,
-    "hiddenGroups",
-    []
-  );
   const [storedPageSize, setStoredPageSize] = useStorageState<number>(
     localStorage,
     "machineListPageSize",
@@ -133,15 +95,6 @@ const MachineList = ({
       ) : null}
       {!headerFormOpen ? <ErrorsNotification errors={machinesErrors} /> : null}
       <VaultNotification />
-      <MachineListControls
-        filter={searchFilter}
-        grouping={grouping}
-        hiddenColumns={hiddenColumns}
-        setFilter={handleSetSearchFilter}
-        setGrouping={handleSetGrouping}
-        setHiddenColumns={setHiddenColumns}
-        setHiddenGroups={setHiddenGroups}
-      />
       <MachineListTable
         callId={callId}
         currentPage={currentPage}
