+++ src/app/machines/views/MachineList/MachineListHeader/MachineListHeader.tsx
@@ -5,21 +5,22 @@ import { useDispatch, useSelector } from "react-redux";
 import { Link, useMatch } from "react-router-dom-v5-compat";
 import { useStorageState } from "react-storage-hooks";
 
+import MachineListControls from "../MachineListControls";
+
 import AddHardwareMenu from "./AddHardwareMenu";
 
-import NodeActionMenu from "app/base/components/NodeActionMenu";
 import MachinesHeader from "app/base/components/node/MachinesHeader";
-import { useSendAnalytics } from "app/base/hooks";
 import type { SetSearchFilter } from "app/base/types";
 import urls from "app/base/urls";
 import MachineHeaderForms from "app/machines/components/MachineHeaderForms";
-import { MachineHeaderViews } from "app/machines/constants";
 import type {
   MachineSidePanelContent,
   MachineSetSidePanelContent,
 } from "app/machines/types";
 import { getHeaderTitle } from "app/machines/utils";
+import { actions as machineActions } from "app/store/machine";
 import machineSelectors from "app/store/machine/selectors";
+import type { FetchGroupKey } from "app/store/machine/types";
 import { FilterMachines, selectedToFilters } from "app/store/machine/utils";
 import {
   useFetchMachineCount,
@@ -28,10 +29,12 @@ import {
 } from "app/store/machine/utils/hooks";
 import { actions as resourcePoolActions } from "app/store/resourcepool";
 import resourcePoolSelectors from "app/store/resourcepool/selectors";
-import { NodeActions } from "app/store/types/node";
-import { getNodeActionTitle } from "app/store/utils";
 
 type Props = {
+  grouping: FetchGroupKey | null;
+  setGrouping: (group: FetchGroupKey | null) => void;
+  setHiddenGroups: (groups: string[]) => void;
+  setHiddenColumns: React.Dispatch<React.SetStateAction<string[]>>;
   sidePanelContent: MachineSidePanelContent | null;
   searchFilter: string;
   setSearchFilter: SetSearchFilter;
@@ -39,6 +42,9 @@ type Props = {
 };
 
 export const MachineListHeader = ({
+  grouping,
+  setGrouping,
+  setHiddenGroups,
   sidePanelContent,
   searchFilter,
   setSearchFilter,
@@ -47,11 +53,6 @@ export const MachineListHeader = ({
   const dispatch = useDispatch();
   const machinesPathMatch = useMatch(urls.machines.index);
   const hasSelection = useHasSelection();
-  const [tagsSeen, setTagsSeen] = useStorageState(
-    localStorage,
-    "machineViewTagsSeen",
-    false
-  );
   const filter = FilterMachines.parseFetchFilters(searchFilter);
   // Get the count of all machines
   const { machineCount: allMachineCount } = useFetchMachineCount();
@@ -59,7 +60,6 @@ export const MachineListHeader = ({
   const { selectedCount, selectedCountLoading } =
     useMachineSelectedCount(filter);
   const selectedMachines = useSelector(machineSelectors.selectedMachines);
-  const sendAnalytics = useSendAnalytics();
 
   // Clear the header when there are no selected machines
   useEffect(() => {
@@ -74,55 +74,42 @@ export const MachineListHeader = ({
 
   const resourcePoolsCount = useSelector(resourcePoolSelectors.count);
 
-  const getTitle = useCallback(
-    (action: NodeActions) => {
-      if (action === NodeActions.TAG) {
-        const title = getNodeActionTitle(action);
-        if (!tagsSeen) {
-          return (
-            <>
-              {title} <i className="p-text--small">(NEW)</i>
-            </>
-          );
-        }
-        return title;
-      }
-      return null;
+  // fallback to "None" if the stored grouping is not valid
+  const [hiddenColumns, setHiddenColumns] = useStorageState<string[]>(
+    localStorage,
+    "machineListHiddenColumns",
+    []
+  );
+
+  const handleSetSearchFilter = useCallback(
+    (filter: string) => {
+      setSearchFilter(filter);
+      // clear selected machines on filters change
+      // we cannot reliably preserve the selected state for groups of machines
+      // as we are only fetching information about a group from the back-end
+      // and the contents of a group may change when different filters are applied
+      dispatch(machineActions.setSelectedMachines(null));
     },
-    [tagsSeen]
+    [dispatch, setSearchFilter]
   );
 
   return (
     <MachinesHeader
       buttons={[
+        <MachineListControls
+          filter={searchFilter}
+          grouping={grouping}
+          hiddenColumns={hiddenColumns}
+          setFilter={handleSetSearchFilter}
+          setGrouping={setGrouping}
+          setHiddenColumns={setHiddenColumns}
+          setHiddenGroups={setHiddenGroups}
+          setSidePanelContent={setSidePanelContent}
+        />,
         <AddHardwareMenu
           disabled={hasSelection}
           key="add-hardware"
           setSidePanelContent={setSidePanelContent}
-        />,
-        <NodeActionMenu
-          alwaysShowLifecycle
-          excludeActions={[NodeActions.IMPORT_IMAGES]}
-          getTitle={getTitle}
-          hasSelection={hasSelection}
-          key="machine-list-action-menu"
-          nodeDisplay="machine"
-          onActionClick={(action) => {
-            if (action === NodeActions.TAG && !tagsSeen) {
-              setTagsSeen(true);
-            }
-            const view = Object.values(MachineHeaderViews).find(
-              ([, actionName]) => actionName === action
-            );
-            if (view) {
-              setSidePanelContent({ view });
-            }
-            sendAnalytics(
-              "Machine list action form",
-              getNodeActionTitle(action),
-              "Open"
-            );
-          }}
         />,
       ]}
       machineCount={allMachineCount}
