+++ src/app/machines/views/MachineList/_index.scss
@@ -132,8 +132,33 @@
     display: inline;
   }
 
-  .machine-list-controls {
-    grid-gap: 0.5rem;
+  .section-header__main-row {
+
+    .section-header__title {
+      margin-bottom: 0;
+    }
+
+    .section-header__buttons {
+      display: flex;
+      align-items: center;
+
+      .machine-list-controls {
+        grid-gap: 0.5rem;
+        align-items: center;
+
+        .p-search-box, .p-form-validation__input, .p-button--link {
+          margin-bottom: 0;
+        }
+
+        .p-button--link {
+          align-self: center;
+        }
+      }
+
+      .p-contextual-menu__toggle {
+        margin-bottom: 0;
+      }
+    }
   }
 
   .hidden-columns-select-wrapper {
@@ -145,6 +170,7 @@
     padding-left: $spv--small;
     text-align: left;
     width: 100%;
+    margin-bottom: 0;
 
     i[class*="p-icon"]:last-child {
       margin-left: 0;
