+++ src/scss/_utilities.scss
@@ -70,6 +70,11 @@
     justify-content: space-between !important;
   }
 
+  .u-flex--center {
+    @extend %display-flex;
+    justify-content: center !important;
+  }
+
   .u-flex--column-align-end {
     @extend %display-flex;
     align-items: flex-end !important;
